[C API] Split IndexIVFFlat to separate files, impl missing functions
